+++ seantis/agencies/tests/test_pdf_export.py
@@ -2,6 +2,7 @@ import mock
 import os
 
 from datetime import datetime, timedelta
+from zExceptions import NotFound
 
 from seantis.agencies import tests
 
@@ -47,7 +48,7 @@ class TestPdfExport(tests.BrowserTestCase):
 
         browser.open('/pdfexport-agencies?force=1')
         self.assertEquals(browser.contents, u'PDFs exported')
-        browser.open('/Plone site.pdf')
+        browser.open('/exported_pdf.pdf')
         self.assertEquals(browser.headers['Content-Type'], 'application/pdf')
         self.assertNotEquals(browser.contents, '')
 
@@ -67,7 +68,7 @@ class TestPdfExport(tests.BrowserTestCase):
         browser = self.new_browser()
         browser.open('/pdfexport-agencies?force=1')
         self.assertEquals(browser.contents, u'PDFs exported')
-        browser.open('/Plone site.pdf')
+        browser.open('/exported_pdf.pdf')
         self.assertEquals(browser.headers['Content-Type'], 'application/pdf')
         self.assertNotEquals(browser.contents, '')
 
@@ -104,3 +105,56 @@ class TestPdfExport(tests.BrowserTestCase):
         now = today + timedelta(hours=12)
         next_run = export_scheduler.get_next_run(now)
         self.assertEqual(next_run, tomorrow + timedelta(minutes=30))
+
+    @mock.patch('kantonzugpdf.ReportZug.get_logo')
+    def test_view_single_pdfs(self, get_logo):
+        get_logo.return_value = None
+
+        self.add_organization(title='a', portrait='portrait')
+        self.add_organization(title='b', path='/a')
+        self.add_organization(title='c', path='/a/b')
+        self.add_organization(title='d', path='/a/b/c')
+        self.add_organization(title='e', path='/a/b/c/d')
+        self.add_organization(title='f', path='/a/b/c/d/e')
+        self.add_organization(title='g', path='/a/b/c/d/e/f')
+        self.add_organization(title='h', path='/a/b/c/d/e/f/g')
+
+        browser = self.new_browser()
+
+        for org in ['a', 'a/b', 'a/b/c', 'a/b/c/d', 'a/b/c/d/e',
+                    'a/b/c/d/e/f', 'a/b/c/d/e/f/g']:
+            self.assertRaises(NotFound, browser.open,
+                              '/organizations/%s/exported_pdf.pdf' % org)
+
+        browser.open('/organizations/a/b/c/d/e/f/g/pdf')
+        browser.open('/organizations/a/b/c/d/e/f/g/exported_pdf.pdf')
+
+        for org in ['a', 'a/b', 'a/b/c', 'a/b/c/d', 'a/b/c/d/e',
+                    'a/b/c/d/e/f']:
+            self.assertRaises(NotFound, browser.open,
+                              '/organizations/%s/exported_pdf.pdf' % org)
+
+        for org in ['a', 'a/b', 'a/b/c', 'a/b/c/d', 'a/b/c/d/e',
+                    'a/b/c/d/e/f']:
+            browser.open('/organizations/%s/pdf' % org)
+            browser.open('/organizations/%s/exported_pdf.pdf' % org)
+
+    @mock.patch('kantonzugpdf.ReportZug.get_logo')
+    def test_view_full_pdf(self, get_logo):
+        get_logo.return_value = None
+
+        self.add_organization(title='a', portrait='portrait')
+        self.add_organization(title='b', path='/a')
+        self.add_organization(title='c', path='/a/b')
+        self.add_organization(title='d', path='/a/b/c')
+        self.add_organization(title='e', path='/a/b/c/d')
+        self.add_organization(title='f', path='/a/b/c/d/e')
+        self.add_organization(title='g', path='/a/b/c/d/e/f')
+        self.add_organization(title='h', path='/a/b/c/d/e/f/g')
+
+        browser = self.new_browser()
+
+        self.assertRaises(NotFound, browser.open, '/exported_pdf.pdf')
+
+        browser.open('/pdf')
+        browser.open('/exported_pdf.pdf')